A couple of fixes to make TLV compile. added tlvconfig_test to regressions.
git-svn-id: http://svn.code.sf.net/p/retroshare/code/trunk@264 b45a01b8-16f6-495d-af2f-9b41ad6348cc

+++ libretroshare/src/serialiser/rstlvtypes.h
@@ -156,7 +156,7 @@ virtual bool     SetTlv(void *data, uint32_t size, uint32_t *offset); /* seriali
 virtual bool     GetTlv(void *data, uint32_t size, uint32_t *offset); /* deserialise */
 virtual std::ostream &print(std::ostream &out, uint16_t indent);
 
-	std::list<uint8_t> ids; /* Mandatory */
+	std::list<uint32_t> ids; /* Mandatory */
 };
